backend/requirements/development.txt
-r base.txt

# Development tools
ipython>=8.20.0
django-debug-toolbar>=4.2.0
